util.c/h -> utils.c/h
