+++ userspace/testapp/testapp.c
@@ -35,7 +35,7 @@
 #include <stdlib.h>
 #include "tests/ipc.h"
 #include "debug.h"
-#include "util.h"
+#include "utils.h"
 
 #define MAP_BUFFER_SIZE 16384
